backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==1.3.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
httpx==0.27.0
